refactor: 8주차 pr review 반영
