+++ src/components/modal/index.js
@@ -1,0 +1,4 @@
+import Modal from "components/modal/modal";
+import BaseModal, { ModalType } from "components/modal/BaseModal";
+
+export { Modal, BaseModal, ModalType };
