+++ src/components/modal/modalType/DeleteModal.js
@@ -24,7 +24,7 @@ const DeleteModal = (props) => {
       default:
         break;
     }
-  }, []);
+  }, [modalType]);
 
   return (
     <>
